Remove extra event service wrapper
